+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 All notable changes to this project will be documented in this file.
  
 ## [Unreleased]
+#### Fixed
+- Fixed the bug where spaces weren't encoded in entitlements
 
 ## [v3.9.0]
 #### Added